moved the get_baseurl function to utils to generate image download links with proper protocol (https)
